add service + controller Volontario

+++ animali/animali/src/main/java/it/adozioni/animali/Repository/VolontarioRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface VolontarioRepository extends JpaRepository<Volontario, Long> {
+public interface VolontarioRepository extends JpaRepository<Volontario, Integer> {
 
     // 🔹 1 - Cerca per nome
     List<Volontario> findByNome(String nome);
